move calcSummary into util folder to make it a common utility
